add Repository class to handle data

+++ app/src/main/java/com/zerox/ticketmanager/data/model/database/dao/UserDao.kt
@@ -13,5 +13,5 @@ interface UserDao {
     suspend fun getUserById(id:Int):UserEntitiy
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
-    suspend fun insertUser(user : UserEntitiy):Long
+    suspend fun insertUser(user : UserEntitiy):Int
 }
